refact: 일부 Enum 위치 변경
